refactor Category Management
