+++ src/main/java/pl/creative/rental_server/ItemTemplateManagement/ItemTemplateApi.java
@@ -1,15 +1,15 @@
-package pl.creative.rental_server.CategoryMagagement;
+package pl.creative.rental_server.ItemTemplateManagement;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.creative.rental_server.CategoryMagagement.dto.FillItemTemplateDto;
-import pl.creative.rental_server.CategoryMagagement.dto.GetItemTemplateDto;
+import pl.creative.rental_server.ItemTemplateManagement.dto.FillItemTemplateDto;
+import pl.creative.rental_server.ItemTemplateManagement.dto.GetItemTemplateDto;
 
 import javax.validation.Valid;
 import java.util.List;
 
-@RequestMapping("/category")
+@RequestMapping("/itemTemplate")
 public interface ItemTemplateApi {
     @GetMapping()
     @ResponseStatus(HttpStatus.OK)
